refactor SwaggerWriter to separate package
